Fix up partially constructed submodules.
Fix submodules on exit as well as entry. This will ensure that gosub
sync no longer leave "semi-submodules" lying around.

It is still helpful to fix submodules on entry too to ensure a good
starting point for syncing. For example, if the user runs an old
version of gosub, they may end up with "semi-submodules" which need
fixing up.

Note that the submodule's git repository is not deleted, although
its path is temporarily removed from the git index.

Signed-off-by: Glyn Normington <gnormington@pivotal.io>